+++ CHANGELOG.md
@@ -62,6 +62,10 @@
   **Rationale**: Python 3.9 is EOL since October 2025.
   Python 3.9 only accounts for less than 0.1% of downloads.
 
+- Removed `format_common_iso()` and `parse_common_iso()` methods.
+  Use `format_iso()` and `parse_iso()` instead.
+  These have been deprecated since 0.9.0.
+
 - `TimeDelta.in_hours()`, `.in_minutes()`, `.in_seconds()`,
   `.in_milliseconds()`, `.in_microseconds()`, `.in_nanoseconds()`,
   `.in_days_of_24h()`, and `.in_hrs_mins_secs_nanos()` are